fix: overflow in last page of "Latest Patchsets from"
In the "Latest Patchsets from" screen, if, for a given target list,
there are, say, 40 patchsets in total, and the page size is 25, the last
page (the second one) will have 15 patchsets. In this case, if the
bottom patchset is selected (the 15th one) and the user continues
feeding input to select the below patchset, the app "will select" the
16th, then the 17th, and so on until the 25th of this page "is
selected". This causes panics, if the user tries to access these invalid
patchsets.

To fix this, use the actual number of patchsets processed instead the
page number times the page size to check if we can select the below
patchset.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>
Co-developed-by: OJarrisonn <j.h.m.t.v.10@gmail.com>