fix: overflow in "Latest Patchsets from"
In a recent commit [1], I fixed the overflow that happened on the last
page of a target mailing list, in which the app would try to select
non-existent patchsets. However, by fixing this, I introduced a possible
overflow on all other pages, as the application only checked if the next
patch number was bigger than the total number of patchsets processed,
ignoring if it was on the next page.

Fix this by also checking if the next patchset is, besides existent, in
the current page.

[1]: 012828367846a9ffa4648f09e5f9bd33fb56901b

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>