feat: add build and unit testing CI workflow
The project lacks CI workflows that cover the process of checking for
compilation problems with `cargo build`. It also lacks one that cover
unit testing with `cargo test`.

Considering that there is no point in unit testing the state of the repo
if it fails in building, add a single workflow that (sequentially)
checks for compilation problems, and runs the unit tests suites.

Note that this workflow is heavily inspired by the recently added
`format_and_lint.yml`.

Closes: #10

Thanks-to: Lorenzo Bertin Salvador <lorenzobs@usp.br>
Thanks-to: OJarrisonn <j.h.m.t.v.10@gmail.com>
Thanks-to: Thiago Duvanel <thiago.duvanel@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>