feat/refactor: Replace FakeLoreAPIClient with automocked BlockingLoreAPIClient
This commit adds the "automock" library to Cargo.toml, which is used to
automatically mock traits and structs, avoiding the need for manual
implementation of each trait function just for testing. Besides that, this
commit also replaces the FakeLoreAPIClient used in src/lore_session/tests.rs
with an "automock" for BlockingLoreAPIClient. This way, each test can have
an explicit expectation for each of the called functions, facilitating
test implementation and improving test quality overall.

Closes: #34

[Maintainer edits]
- Change commit message subject to fit feat/refactor/fix pattern.

Signed-off-by: Lorenzo Bertin Salvador <lorenzobs@usp.br>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>