refactor: define editable configurations
There are some `patch-hub` configurations that just don't make sense to
expose to the user. For example, instead of allowing the user to edit
the paths of bookmarked patchsets, logs, etc., it is cleaner to let the
user set the data directory that will be the common base for all of
those.

In this sense, define page size, cache dir, data dir, and `git send
email` option as the 4 editable configurations and implement setters for
those.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>