feat: add renderers for patch preview
Currently, patch previewing has no styling, coloring or anything that
makes it really readable and, honestly, useful. We intended to implement
in-house ways to deal with this, but we found that we can leverage
external tools for this work, while still having the current behavior
(denominated "default") as a fallback.

Add two renderers for patches: `bat` and `delta`. This might let the
user customize to some level the experience using patch-hub having
a even better visualization of the patch content.

[Maintainer edits]
- Squash commit "refactor: rename patchset_details to render_patchset"
- Squash commit "feat: include the patch_renderer option in sample
  config"
- Squash commit "fix: rendering error message said bat but it's for any
  previewer"
- Rephrase commit message subject
- Expand commit message body

Signed-off-by: OJarrisonn <j.h.m.t.v.10@gmail.com>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>