test: add suite for patch renderers
In recent commit, we added support for external renderer for the patch
previewing. However, they weren't being tested.

Although unit testing won't cover all scenarios, but add a test suite
for (at least) checking if the rendering occurs as expected. Note that
this suite is tagged to be ignored labeled as "optional-dependency",
because it will fail in the setup doesn't have the renderers installed.

[Maintainer edits]
- Make definition of patch sample without tabs at the start. This made
  the renderers "ignore" the rendering and caused silent fails
- Remove redundant definitions of the patch sample
- Add sample files with the expected rendering for each of the three
  renderer options
- Rephrase commit message subject
- Add commit message body

Signed-off-by: OJarrisonn <j.h.m.t.v.10@gmail.com>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>