refactor: handler modularization
The controller component is totally defined as the function `run_app` in
`src/main.rs`. Besides the fact that using `main.rs` to do more work
than setup, start the application, and teardown not being a good idea,
the whole component is a monolithical mess.

To continue the effort of refactoring `patch-hub`, create a dedicated
module for the controller component called `handler` to effectively
drive the application. We took the opportunity to also break down the
monolith into a logic handler and a key handler, and create sub-modules
for every screen.

Helps: #7

[Maintainer edits]
- Rephrase commit message subject
- Add more context to commit message body
- Add "Helps" tag

Signed-off-by: Thiago Duvanel <thiago.duvanel@usp.br>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>