fix: add `PatchRenderer` as case of `TryFrom EditableConfig` to avoid panic
Probably when solving merge conflicts of commit 7b11da3 [1],
I accidently must have deleted the case for `PatchRenderer` when
converting an index to a variant of the enum `EditableConfig`, done by
the late function `from_integer()`. This caused a panic when accessing
the `EditConfig` screen.

To avoid this panic, add `PatchRenderer` as equivalent to the uint 4,
like it was in `from_integer()`.

[1]: refactor: use idiomatic `TryFrom` rather than `from_integer` for `EditableConfig`

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>