fix: save config file in first `patch_hub` run
When `patch-hub` doesn't detect a config file in the system, it tries to
save one with the default values. The method that handles the saving is
`Config::save_patch_hub_config()`. However, this method assumes that the
parent directory of the config path exists, so its saving can silently
fail. By the logic of the code, this will surely happen in the first run
of `patch-hub`.

To avoid this error, assure we are creating the parent directory in this
method before trying to save the config file.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>