feat: introduce static loading screen w/ `loading_screen!` macro
Delayed actions are at the core of `patch-hub`. Currently, we don't have
any notification of loading, which is awful for a handful of reasons.

Introduce static loading with macro `loading_screen!`, which spawns
a thread to run in parallel and render the static loading notification,
while the worker thread runs the delayed task.

Note that we need to take the ownership of the terminal variable to move
it to the thread, so we had to adapt the architecture to pass this
around. Also, it is important to note that some potential problems about
panicking/returning earlier, but through tests we concluded that those
are either impossible to introduce explictly, or when happen (in case of
panicking) the whole execution finishes.

[Maintainer edits]
- Squashed all commits
- Remove introduction of `.vscode/` to `.gitignore`
- Fix merge conflicts
- Rephrase commit message
- Add loading for fetching next page and downloading patchset

Signed-off-by: OJarrisonn <j.h.m.t.v.10@gmail.com>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>