feat: add horizontal panning to patch preview
Patch previewing had scrolling, but no horizontal panning, so if a line
was too wide, it would be impossible for the user to see the content
that overflew the line width.

In this sense, add horizontal panning by adding a field called
`preview_pan` that is analogical to `preview_scroll_offset` (but for the
horizontal orientation), methods that manipulate this value, and use
those in the handler of "Patchset Details and Actions".

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>