refactor: remove duplicated `as_mut().unwrap()` calls
In handlers of "Patchset Details and Actions" and "Latest Patchsets",
there is a lot of binding a key to a method of the respective types of
these screens. Because of this, there as lot of duplications like

```rust
app.patchset_details_and_actions_state
    .as_mut()
    .unwrap()
    .<specific_method>;
```

To reduce this duplications and, collaterally these files sizes, declare
once a mutable reference to these types at the start of the handlers.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>