feat: add `log_on_error!` macro
Creates a macro for automatically logging if some Result is an Err. The
expression itself remains untouched, so it is still possible to
unwrap/expect, apply the ? operator or ignore the Result. The only
difference is that the error will be logged if it happens. The macro
accepts any LogLevel, but the default is LogLevel::Error.

Helps: #47

Signed-off-by: Lorenzo Bertin Salvador <lorenzobs@usp.br>