refactor: extract `log_on_error!` and `log_gc` to dedicated files
The `logging` module is growing to the point that itself having
submodules with dedicated files is justificated.

In this sense, create a specific directory `app/logging` to house the
`log_on_error!` macro and the old `log_gc` module. We took the
opportunity to rename the `log_gc` module to `garbage_collector` for
cleaness (note that this is acceptable to do in the same commit as we
are only renaming the module name).

[Maintainer edits]
- Rephrase commit message
- Remove `app/logging/mod.rs` and recreate `app/logging.rs`
- Move `Logger` struct and implementation to `app/logging.rs`.
  Collaterally revert the imports `logging::logger` to just `logging`.

Signed-off-by: Lorenzo Bertin Salvador <lorenzobs@usp.br>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>