refactor: rename `details` modules to `details_actions`
To be more idiomatic, the modules for the "Patchset Details and Actions"
screen are called `details`, as in `app::screens::details` and
`handler::details`. However, from a nitpicking perspective, this
shortening is somewhat ambiguous.

Given that adding `_actions` doesn't compromises the idiomacy and
conciseness, whilst making it clearer, append it to these modules.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>