feat: add fullscreen for patch preview
In "Patchset Details and Actions", the layout is divided into (1)
details and actions about the patchset and (2) preview of the patches
from the series. Sometimes the user may want to only see the preview of
patches, and the details and actions may become visual pollution.

To provide for this type of situation, add a fullscreen toggle of the
preview by pressing the `f` key. Note that this is hardcoded, but we can
make it customizable when we begin the effort of customizing keybinds.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>