refactor: finish extracting UI components to dedicated files
On a recent commit [1], we extracted the UI related components of the
screen "Patchset Details and Actions" to the file
`src/ui/details_actions.rs`. This was done as the UI component was (at
the start) completely defined in `src/ui.rs`.

This commit finishes the effort of encapsulating the other UI related
components of the remaining screens ("Bookmarked", "Latest Patchsets",
"Mailing List") and the UI component responsible for rendering the
navigation bar with the mode footer and the keys hints.

[1]: commit SHA 20368a9b3e0beeb870ae9e058c54f432389099d8

[Maintainer edits]
- Extract the respective mode footer text and keys hint to the
  respective UI components
- Add commit message

Signed-off-by: Thiago Duvanel <thiago.duvanel@usp.br>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>