feat: add dynamic spinner for `loading_screen!` macro
The `loading_screen!` macro spawns a thread to take care of rendering
the loading notification, which is static at the moment, while the main
thread runs the delayed action(s).

To notify the user that the app is running and didn't halted, add
a dynamica spinner to the rendered notification. As we don't need to
keep the state of the spinner tied with the actual progress of the
delayed action, we can encapsulate its state entirely in the `crate::ui`
module using a const and a static mut, alongside some unsafe code blocks
(ugh).

Note that we probably will want to encapsulate the loading logic to
a module inside `crate::ui`.

Co-developed-by: Gabriel Alves <alvesgabriel@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>