refactor: extract loading rendering logic to `crate::ui::loading_screen`
The loading logic has particularities like a constant and a static mut
variable along with some unsafe code and a `terminal.draw()` call that
violates the main call at the start of the app loop.

Due to these reasons, extract all the loading rendering logic to
a dedicated module `crate::ui::loading_screen`.

Co-developed-by: Gabriel Alves <alvesgabriel@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>