fix: remove `rustc` and `cargo` versions of issue templates
Inspired by what was done in the parent project of `patch-hub`, the `kw`
project, I added the versions of `rustc` and `cargo` in the templates of
the bug report and improvement issues. This does makes sense, but it
sure enough is a hassle to be "mandatory" (the issuer can ignore them,
of course, but it still imposes those).

To make issue reporting more dynamic, remove these fields from the
template.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>