fix: replace `reqwest` with `ureq` for consistent net requests
Replace the `reqwest` crate with `ureq`. These are used for making HTTP
requests, but this replacemente addresses the reported issue of
unreasonable long (and flaky) request latencies, improving the overall
performance of our application.  Also adds a user-agent and a default
"Accept" header.

Note that this also has the collateral effect of making `patch-hub`
leaner, as using `reqwest` was the same as using a cannon to kill a fly,
because, underneath, it used things like `tokio`, which are completely
unnecessary for our purposes of making simple blocking HTTP GET
requests.

Closes: #3

[Maintainer edits]
- Add collateral effects in commit message
- Remove `ClientError::UnexpectedResponse` variant as it became obsolete
  with `ureq` + `ClientError::FromUreq`

Signed-off-by: Rafael Passos <rafael@rcpassos.me>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Tested-by: Lorenzo Bertin Salvador <lorenzobs@usp.br>
Tested-by: OJarrisonn <j.h.m.t.v.10@gmail.com>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>