refactor: use `LazyLock` to compile `lore::lore_session` regexes once
`patch-hub` uses a lot of regexes for processing in many contexts. The
most critical (the ones about fetching patchsets from a list) are
implemented as singletons, but other aren't which can lead to some
performance problems as these are created at every invocation. In this
sense, add the crate `LazyLock` to lazy initialize these instances.

[Maintainer edits]
- Rephrase commit message subject
- Add commit message body

Signed-off-by: auyer <rafael@rcpassos.me>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>