feat: standardize loading screen for opening patchset
Both when consulting the flow of patchsets from a target list or when
navigating through the bookmarked patches, we can open a patchset which
gets us to the screen "Patchset Details and Actions". In the former,
there is a notification of "Downloading patchset", while in the latter
it doesn't.

In this sense, standardize the use of a loading screen with the
notification "Loading patchset" for both cases.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>