feat: add reply w/ `Reviewed-by` to subset of patches
Implement the possibility to only reply to a subset of patches from
a patchset with the `Reviewed-by` tag. Previously, it was only possible
to reply to every patch in the series.

Collaterally, we had to add a field named `patches_to_reply`, which is
a `Vec<usize>`, to `PatchsetDetailsAndActionsState` to keep track of
which patches from the subset to reply with the tag. We also took the
opportunity to make `reviewed_patchsets` from `App`
a `HashMap<String,HashSet<usize>>` instead of
a `HashMap<String,Vec<usize>>`, to ensure uniqueness of patch numbers.

It is important to notice that this commit doesn't implement a reply to
all button, neither updates the UI component for this new changes.

Closes: #78

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>