refactor: remove suffix `State` from screens types
`App` is the macro type that represents the overall state of the
application, while each screen has a respective field in it. These types
had redundant suffixes `State` that just made the names longer (same
goes for the fields in `App`).

With that in mind, remove these suffixes from the screen types, their
respective fields in `App`, and variable names.

Closes: #92

[Maintainers edits]
- Rephrase commit message subject
- Rephrase commit message body

Signed-off-by: Gabriel Alves <alves.gabriel@novaenergia.com.br>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>