feat: make patchset apply action interactive and with feedback
The previous command introduced all the logic to handle the apply of
a patchset into a target kernel tree. However, it lacked any
interactiveness and feedback for the user.

In this sense, add a button "apply" to the "Actions" tab of the
"Patchset Details and Actions" screen, and make it run like the other
actions, i.e., after the user toggles it by hitting (in this case) the
`a` key, the `ENTER` key needs to be hit to actually run the action.

Moreover, implement a new `PopUp` type, named `InfoPopUp`, and use it to
display the result of the patchset apply action.

[Maintainer edits]
- Rephrase commit message
- Switch order of "apply" button with "reviewed-by"
- Implement and use `InfoPopUp` to notify user about result of patchset
  apply action

Signed-off-by: OJarrisonn <j.h.m.t.v.10@gmail.com>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>