fix: handle logical events before drawing UI to avoid stale frames
In a recent commit [1], I removed the auto-refreshing of the UI due to
it consuming high CPU resources unnecessarily for the nature of
`patch-hub`. I stick with this opinion, and don't see in the short-term
period a justification to enable it.

However, there is an "edge-case" which slipped by, which was when there
was events that don't required inputs (handled by the function
`handler::logic_handle()`). This function runs after `terminal.draw()`,
the main rendering point of our UI. If `logic_handle()` draws
a a loading screen, the loading screen will be freezed until the user
inputs something. This misleads users to think that there has been
a silent crash. To reproduce this, just delete `patch-hub` mailing lists
JSON (probably at `~/.local/share/patch-hub/mailing_lists.json`), and
launch the tool.

To fix this, in each iteration of the application loop, first handle
logical events, then draw the UI.

*Note*: Due the severity of the bug, this commit is being pushed
directly to the upstream.

[1]: 841a6faa1fd0ec399642792086bbe26f8e1cc9bb

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>