feat: show list of patches that already have been `Reviewed-by`
In the last commit, we updated the UI to reflect the new functionality
of replying to a subset of patches from a series w/ the "Reviewed-by"
tag.

Make the line "Reviewed-by:" in the "Details" tab, display information
not only about the staged patches, but the ones recorded as already
replied with the tag, differentiating them by the color and the
appended `*`.

Following the example of the last commit, if a patches 1, 2, and 7 are
staged, and 0, 3, and 10 have already been replied with the tag, the
line will display as

```
Reviewed-by: (0, 1*, 2*, 3, 7*, 10)
```

Closes: #78

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>