feat: create the PopUp trait
The PopUp trait defines the model for how patch-hub handles PopUps

The app holds an `Option` for the current popup being displayed. The app
renders the popup on top of every component. Also when a popup is being
displayed it hijacks the keyevents (except for the ESC key which always
closes the popup)

A popup implementing PopUp must:

- define its size so the app can produce a centralized Rect for it
- define its rendering method which receives a Rect of the given size
- define how it handles key events

Signed-off-by: OJarrisonn <j.h.m.t.v.10@gmail.com>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>