feat: patch cover colors
This uses similar design for patch rendering. The cover of a patch is
everything before the `---` line that separates the email part of a
patch and the actual diff.

Currently, only `bat` was configured as an option besides the `default`

This also includes a config option for the cover renderer

Signed-off-by: OJarrisonn <j.h.m.t.v.10@gmail.com>