refactor: remove unnecessary `current_screen` arg from `App::init_details_actions`
The method `App::init_details_actions`, used to initialize an instance
of the screen "Patchset Details and Actions", has an argument
`current_screen` to resolve if the patchset to be opened is from the
latest or bookmarked patchsets. However, this can be resolved by just
checking the field `App::current_screen`, so remove this unnecessary
argument.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>