fix: clear UI before rendering anything for sanitizing
There is a flaky bug that, sometimes, the background color of texts
renders differs from the background without rendered texts, as if the
background color gets "stale". This bug is considered flaky, as no
contributor/maintainer could reproduce, but they are consistent in
a user setup.

To try solving this issue, clear the whole UI using the `ratatui`
widget `Clear`, which does exactly what the name entails, before
rendering anything.

[Maintainer edits]
- Rephrase commit message
- Make the clear global, in the sense of clearing before rendering
  anything, not only for the "Mailing List Selection" screen

Signed-off-by: Thiago Duvanel <thiago.duvanel@usp.br>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>