feat: add stats about patch reply trailers
The `b4` tool [1] applies reply trailers when fecthing patches, i.e., if
a patch has already been replied with the tags "Reviewed-by",
"Tested-by", etc. (referred by `b4` as "code-review trailers") when the
the tool consults the thread, it adds them to the returned .mbox. This
enables us to present useful information to users about the review
status of the patch.

To make the "Patchset Details and Actions" screen more useful, extract
these trailers when initializing the screen and present them in the
"Details" tab as a line like

  Reviewed-by: 2 | Tested-by: 0 | Acked-by: 1

[1]: https://b4.docs.kernel.org/en/latest/

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>