[REVIEW] feat: add config options for upcoming patchset apply feature
Create config options needed to apply a patchset to a kernel tree. The
configurations added are:

- `kernel_trees`: map of arbitrary identifiers to kernel trees
- `target_kernel_tree`: identifier of active kernel tree
- `git_am_options`: additional options to the `git am` command
- `git_am_branch_prefix`: prefix of branches when applying patchset

It is important to note that a "kernel tree" in this context of applying
patches means a path to a git repository with a Linux kernel source code
paired with a base branch. This is represented by the type `KernelTree`.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>