[REVIEW] feat: add patchset apply action
Introduce an action to the "Patchset Details and Actions" screen to
apply a patchset to a target kernel tree, i.e., apply the patches in
order using `git am` using a Linux kernel repo and a branch as base.

This action should be a core feature of `patch-hub` and it will be much
more refined and expanded. However, some effort was put into handling
virtually all setup errors (no configs set, invalid branch, rebase in
progress, etc.) and into clean up after the fact (abort the `git am` in
case if fails, and checkout back to the original branch either way).

Currently, there is only a binding of the `a` key to the action, but no
UI drawn. The function binded with the key always returns a string with
information about the status of the action, which can then easilly be
displayed to the user in a `PopUp` type.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>