style: change "Patch Hub" instances to "patch-hub"
At the start, I styled references of `patch-hub` inside of it as "Patch
Hub", which, IMHO, doesn't fit (really can't explain further than that).

In this sense, change the style of those instances to "patch-hub". Also,
make the banner at the top of every screen that says "patch-hub" in
itallic, just bold.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>