feat: implement logger actor
Here i've created the implementation of the logger actor. This actor is
responsible for logging messages.

The `Logger` struct is a simple refactor of the old
`app::logging::Logger` it is responsible for storing messages that will
be later printed to the stdout and also to provide methods such as `spawn`

`spawn` is responsible for creating a dedicated task that will handle
all the messages received from all the transmitters.

The `LoggerTx` is a transmitter that is responsible by sending messages
down to the running actor. This is what should be used accross the code
base to communicate with the actor.

The `MockLoggerTx` is a mock version of this logger which does nothing
at all.

Finally but not least, the `LoggerActor` trait which unifies the
interface of `LoggerTx` and `MockLoggerTx`. This permits the swaping of
the loggers for testing scenarios.

Signed-off-by: OJarrisonn <j.h.m.t.v.10@gmail.com>