fix: trim trailing `--` from patches before previewing
Patches commonly have a signature delimiter `--` at the end (followed by
the git version, sometimes). This is necessary for git to apply patches,
delimit the end of diffs, etc. This is irrelevant to the patch preview
and has a collateral bug because it is interpreted as a deleted line in
the renderers supported (`delta`, `bat`, and `diff-so-fancy`).

In this sense, "clean" the patch before rendering for preview by
removing the `--` line and the following ones.

Fixes: #76

[Maintainer edits]
- Rephrase the commit message subject and body
- Add `Fixes` tag
- Rephrase `clean_patch_for_preview` documentation

Signed-off-by: Ivin Joel Abraham <ivinjabraham@gmail.com>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>