refactor: use the config actor
Adapts the code to use the new config actor instead of the old config
implementation. This mean that several functions became async.

There is a little trick that will be solve later since we cannot use
async functions to draw on the terminal so we pass the `page_size` value
directly instead of gathering it from the `app.config` in the `draw_ui`
function.

Signed-off-by: OJarrisonn <j.h.m.t.v.10@gmail.com>