feat: implement env actor
This actor is a way to improve patch-hub testing suit. It's just a
wrapper to `std::env` to let us manipulate environment variables. But it
also has a mock variant which won't touch the real environment variables
but use a simple `HashMap` instead

With this, the later config tests can be run in parallel

Signed-off-by: OJarrisonn <j.h.m.t.v.10@gmail.com>