feat: disable gitconfig reading by Delta
By default, Delta loads the configuration from the Git local or
global files. This behavior happens even in a context not related
to Git outputs, e.g. using Delta as a fancy replacement for Git.

This may leave an unexpected behavior for patch-hub, as it will
look like different on different setups. Disable the gitconfig
loading using --no-gitconfig, so we can tune the Delta output in
a way that matches better the output that we expect in patch-hub.

Signed-off-by: Lucas Oshiro <lucasseikioshiro@gmail.com>