fix: `cargo install` failing to build when using `ureq v3.0.9`
This commit locks the version of `ureq` to `v3.0.0-rc2`, a tested,
working build.

Version strings specified in `Cargo.toml` represent a range from the
specified version to the next `minor` upgrade according to `SemVer`.
And since `cargo install` does not use `Cargo.lock` to build the
project, it will use a patched version of `ureq` which changes the
function signature of `ureq::Agent::user-agent`.