fix: ensure loading_screen! generated thread joins before exiting patch-hub
This commit fixes the cursor becoming invisible when crash occur during loading.
The problem was caused by not joining the thread generated by loading_screen!
before handling the error (e.g exiting patch-hub).
The solution was achieved by the following changed:
1. Made the closure that's passed to the loading_screen! macro return a Result
instead of (), enabling proper handling of the result after joining the loading
thread.
2. Changed the closures to avoid propagating errors using the ? operator, as
in Rust, "?" and "return" are applied to the outer function instead of the
closure itself.

Closes: #74

Signed-off-by: Lorenzo Bertin Salvador <lorenzobs@usp.br>