feat: create proc_macro to simplify Config attributes deserialization
This commit adds a procedural macro to patch-hub! The intention is
to simplify the Config struct by avoiding to add a #[serde(default = "")]
for each of its attributes and implement the default function.
The proc_macro supports all kinds of structs and should preserve
all other struct attributes, derivations and general behavior.
This completely solves #118.

Signed-off-by: Lorenzo Bertin Salvador <lorenzobs@usp.br>