refactor(config-tests): use untracked temporary sample config file
In the test suite of `App::Config`, we make use of a sample config file
located at `src/test_samples/app/config/config.json`. In solving an
issue about overwriting (issue #118), there is a side-effect of altering
this file, which is tracked by git, resulting in having to potentially
commit redundant changes related to formatting.

To solve this without compromising the solution to issue #118, make the
test suite use an untracked temporary file with the same contents as the
tracked file, but that can be manipulated however we wish.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>